Move view state to new module.
